Improve TextInterpolator layout when rendering multiword RTL text
The offset being used to render text animations for multiword RTL text
was incorrectly being applied on a per-word basis instead of a per-line
basis. This caused each word to functionally overlap as the word offset
would exactly cancel out the offset of the first glyph in a given word.

This change also includes some consistency improvements for clock text
views to ensure they only ever render as one line. We do this by using
maxLines to ensure the layout will not insert a newline, and rounding
up when measuring the view size.

Finally, it also includes a hack which mostly accounts for a very large
x translation that's being applied to the canvas. Visually this approach
is close enough but incorrect. b/450350391 covers follow up to replace
this hack with an accurate solution.

Bug: 450449281
Bug: 443704172
Test: Checked each clock on device
Flag: NONE - Minor RTL Text rendering fix
Change-Id: I92ba5383c9c0d6f827e26954c9d94a79625868e5