Split the encrypted_client_hello_configuration into two
This chagne makes two files of flag
encrypted_client_hello_configuration. In the conscrypt_nsc_off, the flag
is exported, and in the conscrypt_nsc_on, the flag is not exported. The
reason is when the build flag is off the flag won't be used to guard any
API.

Bug: 397646538
Test: build-flag set RELEASE_CONSCRYPT_NSC true && m
Flag: build.release_conscrypt_nsc
Flag: android.security.conscrypt_network_security_config
Change-Id: Iaf16f5c3880f1ff92d6fa131d064ebbfd2381133