fix(eventDispatcher): use the event's eventTime as downTime if no prior DOWN
When EventDispatcher#sendMotionEvent for non-DOWN events, it would use
the cached down time in TouchState as the new downTime for transformed
events. However for HOVER events, the TouchState may not cache the down
time and return 0, then the sent events would have downTime 0.

After ag/36205988 is merged, the input verifier would check the down
time of the injected events. The downTime 0 causes the failures when
check `event.down_time_nanos < old_down_time_nanos`.

Therefore, when there is no cached down time in TouchState, we should
not use 0 as the downTime, but should use the original eventTime or
downTime of the event. Here we choose the eventTime to be consistent
with existing logic.

Add some test cases in EventDispatcherTest to ensure we don't use the
time from TouchState when it does not cache any down time.

Besides, refer to the comment in this CL, we also remove the downtime
cleanup in TouchState to be consistent with real mouse behavior.

Bug: 405380658
Flag: EXEMPT BUGFIX
Test: atest android.accessibilityservice.cts
      atest EventDispatcherTest
      manually check with magnification/talkback
Change-Id: I0230d408393c709937d2bc53d7c0bb3613900942