Manage top_app binding for inference service
This change introduces a mechanism to maintain a high-priority service binding (`BIND_SCHEDULE_LIKE_TOP_APP`) to the inference service when applications in the foreground are actively using it.

The `OnDeviceIntelligenceManagerService` now tracks ongoing inference jobs per UID. When a UID initiates an inference request, its importance is checked. If the UID is in the foreground or higher, a shared high-priority binding is established. This binding is maintained as long as at least one foreground UID has an active job. The binding is released when all foreground UIDs finish their jobs or transition to a background state.

An `OnUidImportanceListener` is used to react to UID importance changes and update the binding status accordingly.

Change-Id: Ia2b44430decc667589736af398a633cb71b17a5b
Bug: 424729565
Flag: EXEMPT BUGFIX