Fix show taps setting after screen record
Currently we only update the existing show_taps setting if the choice
for recording differs from the current setting, but always restore it
for the case where the recording ends successfully (i.e. the phone isn't
rebooted in the middle). This means that if the setting was already on
(and the recording is taken with show_taps on), it will get reset at the
end of the recording.

We should just always check whether the "UPDATE_SHOW_TAPS" value was set
to true, and only reset for that case.

Bug: 450760053
Fix: 450760053
Test: manual (via taking a recording with show_taps already on)
Flag: EXEMPT BUGFIX

Change-Id: I5752368caf42bdc045608a7e5ae0e5e4f8d02374