Improve "Too many receivers" exception debuggability.
When an application registers more than the allowed number of broadcast
receivers (currently 1000), the system throws an IllegalStateException.
However, the current exception message "Too many receivers" is not
very helpful for clients to identify the source of the issue.

This change enhances the exception message to include a detailed summary
of the top 10 most frequently registered broadcast actions and receiver
classes for the application. This additional information will help
developers quickly diagnose and fix issues related to excessive
receiver registration.

Bug: 451599876
Flag: EXEMPT DEBUG
Test: atest com.android.server.am.BroadcastControllerTest
Change-Id: I52248a98c7c75142e4e049b9ee03a5c533f3206a