Handle and filter non-touchscreen motion events in DragDetector
Currently, only touchscreen motion events are handled in DragDetector
before passing them to MotionEventHandler.
This is to not skip move events smaller than the slop threshold when the
event source is non-touchscreen, e.g., mouse or touchpad, based on the
assumption that slops matter only for touchscreen events.

However, we should actually delay the start of a drag until the cursor
is moved large enough even with the mouse/touchpad; otherwise a minute
move between a button down event and a button up event is interpreted as
a drag, invalidating the click action and causing an annoying UX.

With this CL, non-touchscreen events are handled in the same way as
touchscreen events, so that small moves from non-touchscreen sources are
skipped/filtered in the same way as small touchscreen moves, except that
the hold-to-drag logic for the app handle (introduced by ag/29448505) is
not applied to non-touchscreen events.

Note that, regardless of this change, DragDetector does not gracefully
handle motion events from mixed sources (e.g. press and hold the mouse
button and drag with the touchscreen).
This issue about mixed sources will be addressed separately, but this CL
should make it easier to address the issue by handling all motion events
in an integrated way.

Bug: 435051530
Test: atest WMShellUnitTests:DragDetectorTest
Test: With the flag enabled, a small drag action with the mouse/touchpad
      on the minimize/maximize/close app window button does not initiate
      an actual drag, and releasing the mouse/touchpad button from that
      state is treated as a click.
Flag: com.android.window.flags.handle_non_touchscreen_events_in_drag_detector

Change-Id: I4137cb6d8ba4c14099470175473f7c3f83487a1b