[SB] Make chips large enough for touch accessibility.
Increasing the height of the start side icons in status_bar.xml is a
partial revert of I0b20aaf075a379803e43f5e8cd7f9de5c939b332. So, most of
this CL is implementing a different solution for the hover height on the
start side container where we force-set it to be the icon height.

Flag: EXEMPT BUGFIX
Bug: 437078083
Fixes: 446113663

Test: atest StatusOverlayHoverListenerTest PhoneStatusBarViewControllerTest
Test: Verify chip size is 48dp (but background is smaller)
Test: Verify hover over left side of status bar is the same as before,
and adjusts on display size

Change-Id: I55847d76dca3aa1e55168a06921d6b564d9d4872