Fix issues with user switching during going away logic.
a) Going away transitions can be initiated by WM (rather than requested by SysUI) when activities are started with FLAG_DISMISS_KEYGUARD. If the first going away transition after boot is WM-initiated, we never set the user ID, so it was left at -1 and mismatched. This causes the first WM-initiated going away transitions to be ignored (which is common in tests).
b) [flexiglass] Attempt device entry in the insecure case, since startDismissKeyguardTransition is flagged off in Flexiglass.

Fixes: 450253023
Test: atest WindowManagerLockscreenVisibilityManagerTest
Test: atest KeyguardTests
Flag: com.android.systemui.scene_container
Change-Id: Ic1dca352963112e75ca649127da0384e1a1f26b9