Create test ProcessRecords with a Builder
There are currently several helper methods to generate a ProcessRecord
for BroadcastQueueTest. This style of object creation makes it difficult
to add a new argument or dimension to affect the object creation. A
Builder pattern makes it much easier to update and add the object
configuration.

BYPASS_RECITATION_REASON=Flagged code is already existed and was just
moved into a builder (plus no AI was used to generate this code).

Flag: EXEMPT TEST_ONLY
Bug: 315716508
Test: atest BroadcastQueueTest
Change-Id: Ief91b967c0c6c7ffd7ec4db1f68b20f02e6cce0c