Handle Default & Recently Used Folder Selection
In the dropdown to allow users to choose a custom folder for them to
save their screenshots, we want to offer/display two quick select options: 1) Their recently used folder 2) The default (pictures/screenshots) folder. To do this, we want to also add a new data storage within our existing datastore to track whether the user has their screenshot save location to be the default folder or a custom folder. The main reason we are doing this is because the user may want to switch to the default folder, but we still want to retain the recently used folder which they can select back to, thus the need to store both the custom save location URI and its active status.

Bug: 452410311
Test: LargeScreenCaptureParametersRepositoryTest,
LargeScreenCaptureParametersInteractorTest, PreCaptureToolbarViewModelTest
Flag: com.android.systemui.large_screen_screenshot_save_location

Change-Id: I82d3115abac8000f11f0c9f4471a356f99c34294