[Ravenwood] Experimental support for SettingProvider
In order to support TestableSettingsProvider usage in sysui tests, we
experimentally enable the entire content provider/resolver subsystem and
register a stub IContentProvider for the Settings authority.

By doing so, the entire Settings class and Content{Provider/Resolver}
classes can be used as is, laying the groundwork for proper content
provider support in the future.

Bug: 319647875
Flag: EXEMPT HOST_MODE
Test: f/b/r/scripts/run-ravenwood-tests.sh SystemUiRavenTests
Change-Id: I2383b781e208fcd952842bf640dbc5b277a1cb37