Isolate PluginInstanceTest from SharedPrefs
PluginInstanceTest was using a live SharedPreferences object, which was
persisting data to device. This meant the test would consistenly pass in
Robolectric, or the first execution on a freshly wiped device. After the
initial run, subsequent runs would load the SharedPreferences data from
disk and fail. This is corrected for by mocking the SharedPreferences
object used by these tests.

Bug: 447830979
Flag: TEST_ONLY
Test: atest PluginInstanceTest --iterations 10
Change-Id: I4d7d523baca3cfb2761123741dc49fd1f7054dad