Do no show the bubble task surface when transition ready
When starting an app to bubble, both the app task and the bubble
root task was collected in the transition, and the bubble root
task surface was reparented to the transition root when the
transition is ready.

The bubble root task was initially set to invisible (alpha=0) in
Transitions#setupStartState, but was set to visible (alpha=1)
in BubbleTransitions. Therefore, the task surface (the starting
window) is shown immediately and causes flicker.

This issue is happening only for the first time starting an app
to bubble after device reboot. Perhaps related to how starting
window renders its first frame somehow.

Bug: 407669465
Test: start an app to bubble after device boot
Test: BubbleBarAnimationHelperTest
Flag: com.android.window.flags.root_task_for_bubble

Change-Id: Iff69ced745694fa107d97b8a71cba853e9064be8