Fix potential deadlock in AccessibilityInputFilter
In certain situations when TalkBack is enabled and mouse keys becomes
disabled, AccessibilityInputFilter can register a pointer motion filter
with the InputManager. This call was made while holding the
AccessibilityManagerService lock, which could lead to a deadlock if the
InputManager was also trying to acquire the same lock.

This change moves the registration of the pointer motion filter to the
main thread's handler, ensuring that the AccessibilityManagerService
lock is not held during the call. Additionally, this change allows a
Handler to be provided in the AccessibilityInputFilter constructor to
ensure tests are deterministic.

Bug: 450741583
Flag: com.android.server.accessibility.enable_magnification_follows_mouse_with_pointer_motion_filter
Test: Added a new test to verify that the AccessibilityManagerService
      lock is not held when registering the pointer motion filter.
      atest AccessibilityInputFilterTest

Change-Id: I6832e29e50acf8322ac60019c25c44861a92ceca