Remove Obsolete NPE Checks
A long long time ago, TelephonyManager would try to
grab a ref to ITelephony, and if it failed, it would
throw a NPE; however, this is super expensive in terms
of CPU, so null pointer checks were added; however, all
the places where NPE was caught were not cleaned up.
This does mean that if there are NPEs that were being
thrown from the other side of the ITelephony binder, they
were also being caught, and they will now propagate to
callers; this would be a behavioral correction because
NPEs from the other side of the ITelehpony binder should
tell the caller that they passed an invalid argument.

As part of this exercise, convert all remaining places
where the packageName and attributionTag might also be
null to null-safe functions. This supports the instances
where a "static" TelephonyManager is created without a
Context object.

Bug: 19828008
Test: atest TelephonyManagerTest
Flag: EXEMPT BUGFIX
Change-Id: I1ee16338f807b638f3281ec5e513e892418799ad