Screen capture: Disable showing pre-capture UI if keyguard showing
When the lockscreen is showing, fullscreen screenshots can be taken
using the keyboard shortcuts or system key. However, the large-screen
pre-capture UI should not show on the lockscreen since it will expose
screen recording capabilities which should not be invoked if the user is
not signed in.

Bug: 453688314
Test: atest ScreenCaptureKeyboardShortcutInteractorTest
Flag: com.android.systemui.large_screen_screencapture
Change-Id: I3693f8c7053665606295df873f1e226f95f5f06e