Fix blocked state handling in ComputerControlSessionImpl
1. Introduce onActivityLaunchRequested callback in VDM
ActivityListener to know about all activity launch requests
and preemtively detect blocked state in ComputerControlSessionImpl
(based on allowed packages and users).
2. As a result, no need for any custom activity policies and no
need for specifying any set of allowed users in VDM.
3. Use onTopActivityChanged to attempt transition to active state
when an activity launch is completed for an allowed package and user.

The intention of this change is to allow everything in the MirrorView
and block everything from agents in the blocked state.

This fixes the following issues with the previous implementation:
1. Disallowed activities (such as permission dialogs, etc) failed
to launch in the MirrorView (as we had custom activity policy, and
GenericWindowPolicyController#canActivityBeLaunched returned false).
2. As the POLICY_TYPE_ACTIVITY was being made POLICY_DEFAULT in the
active state, calling launchApplication with an allowed package added
it as an exemption (to the POLICY_DEFAULT), which means it was being
blocked by GenericWindowPolicyController for subsequent launches.
3. Session remained in blocked state even after display became empty
after removal of activity with secure window.
4. Session blocking logic did not consider set of allowed/disallowed
users.

Bug: 442540670
Test: atest ComputerControlSessionImplTest, atest
GenericWindowPolicyControllerTest
Flag:android.companion.virtualdevice.flags.computer_control_block_input_and_screenshots

Change-Id: I92ad2508523ef56cc11a9d831429436b2fa6f617