Avoid applying start transaction too early with multiple displays
It should be applied by the animation handler, e.g.
  DefaultTransitionHandler#startAnimation {
    for changes {
      new ScreenRotationAnimation(startTransaction)
    }
    startTransaction.apply
  }
Otherwise if there are multiple display changes in a transition,
the first creation of ScreenRotationAnimation will detach the
snapshots of other display (default operation in startTransaction)
and re-attach until creating their ScreenRotationAnimation, which
may cause potential flickering.

Because the transition handler is responsible to apply the start
transaction, the apply in ScreenRotationAnimation's constructor
in unnecessary.

Bug: 450302345
Flag: EXEMPT BUGFIX
Test: Create multiple virtual displays and toggle dark theme.
Change-Id: I0608188fd6355f6a27d17602a479befd9262630d