Add higher level install permission for localhost
If an app has this permission, loopback traffic is permitted regardless
of the permissions held by the other app. The intention is for this to
be a role permission used by only the DEVICE_POLICY_MANAGEMENT role, and
platform signed apps.

Bug: 441592715
Test: compile
Flag: android.permission.flags.use_loopback_interface_permission_enabled
Change-Id: I6cb6add478f979833f44669292a417d6059729c1