Show shared icon when a network is shared in QS line items
On a multi user device, if a network is shared, this change shows a
shared icon in the QS wifi line item. The shared icon will only
be visible if the entry is not connected.

Flag: com.android.systemui.qs_wifi_config
Flag: com.android.systemui.qs_tile_detailed_view
Bug: 406549709
Bug: 448707226
Test: atest InternetAdapterTest
Test: atest InternetDialogDelegateLegacyTest
Test: atest InternetDetailsContentManagerTest
Test: manual test
Change-Id: Id3acf6a54b279744e9970746c68ba815d2f26e03