Allow opt-in drop during display change to WCT
When a transition is directly started by Shell,
Shell sometimes might not be aware of an undergoing display change
in Core; this might leave a window in a weird state if the changes
requested by the new transition were based on the latest display state in Shell

We have allowed dropping transitions if they arrive during display
change before, but this requires TransitionHandlers to be resilient
and ready to handle potential transition aborts. Hence, a better
approach is to make such changes opt-in to be dropped if deemed "risky".

Bug: 450107220
Flag: EXEMPT BUGFIX
Test: atest WMShellUnitTests:PipScheduler
Test: manually repro the steps in the bug
Change-Id: I169be48dda4f9f35d814094181559dd1ee0c7c78