[flexiglass] Add new BiometricUnlockModes
Add new BiometricUnlockModes to differentiate them
from their non-unlocked variants:
 * MODE_ONLY_WAKE_UNLOCKED
 * MODE_NONE_UNLOCKED

When flexiglass isn't enable, these modes have the
same logic in BiometricUnlockController.
 * MODE_ONLY_WAKE[_UNLOCKED] only wakes the device
 * MODE_NONE[_UNLOCKED] does nothing with regard to waking

When flexiglass is enabled, we use these states to determine
whether to update DeviceUnlockedInteractor's unlocked state.
If the device isn't unlocked (MODE_ONLY_WAKE / MODE_NONE), then
do not set the unlockSource. However, if the device was
unlocked (*_UNLOCKED and other biometricUnlockModes), then
we do want to set the unlockSource.

This fixes a bug where all BiometricUnlockStates would
cause the DeviceUnlockedInteractor to incorrectly believe
the device was unlocked. This is NOT true for all the
BiometricUnlockStates. For example these states should
not result in an unlocked state:
 * MODE_NONE
 * MODE_ONLY_WAKE
 * MODE_SHOW_BOUNCER

Test: atest DeviceUnlockedInteractorTest
Test: atest DeviceEntryInteractorTest
Test: atest BiometricsUnlockControllerTest
Flag: com.android.systemui.scene_container
Fixes: 454329345

Change-Id: I72e83d6c012f9a1fc7ae5d4b6776896a6f8f3b78