STL: Default GestureContext from intrinsicDirection 1/2
This property acts as a high-level hint for the overall direction of a
transition. It is intended to be used by triggered animations as a
fallback for `MotionValue`s that require a direction but do not have one
explicitly defined.


This new implementation provides a `GestureContext` if and only if
`transformationSpec.intrinsicDirection` is defined in the transition's
DSL. This allows programmatic animations (like `animateToScene`) to
provide directional context to mechanics, which they previously could
not.

A key consequence is that `gestureContext != null` is no longer a
reliable indicator of a user-driven gesture. Consumers must now use
`isInitiatedByUserInput` for that check (as seen in BouncerContent).

Test: atest SceneTransitionLayoutTest
Test: atest VerticalTactileSurfaceRevealModifierTest
Bug: 392535471
Flag: com.android.systemui.qs_tiles_reveal_animation
Change-Id: Icabe8a5b5cae709ead7ab5e2889a39a96ddc549b