MotionEvent: add more cases to ACTION_HOVER_{ENTER,EXIT} JavaDoc
As well as when a hovering pointer enters or leaves a window or view,
these events can be emitted without any change in the (X,Y) coordinates
of a pointer, either when the pointer goes down or up, or when a stylus
moves in the third dimension. This is important to document.

Bug: 454264090
Test: none
Flag: EXEMPT DOCS_ONLY
Change-Id: I84e7517ad4d04b50e6d954bb3724663a9c3e9651