Add support for fallback character for combining accents
When a combining accent is followed by a Space or itself, we should
output a default character. For some combining accents, we should
fallback to a different character than the combining accent.
e.g. ◌̃+S Space should be ~ and not ˜

So, for cases where we have a better fallback, we should fallback to
that instead of just outputting the combininng accesnt itself.

Bug: 447547183
Flag: EXEMPT BUGFIX
Test: atest KeyCharacterMapTest
Change-Id: Ic5f55b4ce404f39e7fd500d2cc2e2febd8f21fc2