[Flexiglass] Reset Notification alpha after Bouncer to Shade transitions
The alpha flow fading out Notifications was incorrectly part of a flow, that was used for the shade expansion states. This plumbing required the bouncer flow NOT to fire when a hide(BouncerOverlay) transition has finished. This left NSSL alpha in an inconsistent state after a BouncerOverlay -> Gone transition, which left the HUNs with a transparent alpha.

This CL fixes it by moving the BouncerOverlay transition's alpha to its right place, where it can emit(1f) after BouncerOverlay goes hidden.

Fixes: 443054750
Test: [SingleShade] go to Lockscreen ->
      go from BouncerOverlay directly to Gone ->
      receive a HUN -> HUN is displayed
Flag: com.android.systemui.scene_container
Change-Id: I40f832aeea58bba9ff5180c57db48c64fd33988f