Build out the renderer workflow
After this we'll have ContextInsights routed to either one renderer
(if there's a RenderToken), or all of the renderers (if there isn't
a RenderToken). In the second case it asks each renderer if they want
to handle the ContextInsight (we don't bother asking in the first
case). Once we have filters on the renderers this will automatically
use them.

ContextInsights are validated against a ruleset (they must have non-
conflicting RenderTokens and all ContextHints must be properly
signed.

This does not handle transformers yet.

Bug: 434644900
Fixes: 452425186
Test: atest PersonalContextTests
Flag: android.service.personalcontext.enable_personal_context_service
Change-Id: Ibceb1231fd6ed819c947bcc85dd8e8766a6a6964