Catch uncaught IllegalStateException
MediaProvider may throw an IllegalStateException during some race
conditions on the Screenshots folder. The exception is then passed over
the binder and rethrown in the call to ContentResolver.insert. There's
not a whole lot that can be done in this case, and error handling
already seems to be in place in this function, so I think the best we
can do is catch the exception and eat it with a warning log.

Bug: 441090735
Bug: 393608784
Flag: EXEMPT CVE_FIX
Test: Take a few screenshots
Change-Id: I787ea2f2e449b1a88e8f0c33051bd2a5959970ac