Skip synthetic scrolling during transitions
If synthetic scroll is applied during a transition, it results in a
transition state of transitioning from gone to shade that never ends.
This prevents the scene container window from going invisible, which
results in it consuming all touches and blocking app launches. This
change only allows. This change ensures that synthetic scoll is only
applied when the scene is idle and showing notifications.

Fixes: 449982657
Test: attempted to reproduce the bug 50 times unsuccessfully
Flag: com.android.systemui.scene_container
Change-Id: Id16d93555189f31daaa2b5ef6776931d5212c4ff