Don't show adb wifi auth prompt during auto-enable
Problem: During auto-enable, the AdbWifiNetworkMonitor connects to
a trusted network, and enables adb wifi. During the enablement flow, the
getCurrentWifiApInfo can return past or future value which isn't a
trusted network. This causes the adb wifi authorization dialog to show.

Solution: The AdbWifiNetworkMonitor sends a message to
AdbDebuggingHandler to indicate that the adb wifi enable request is
auto-enable. The AdbDebuggingHandler uses that information to not show
the adb wifi auth prompt.

Test: AdbWifiNetworkMonitor
Bug: 452595025
Flag: EXEMPT BUGFIX

Change-Id: Icdf4ef92719fde8a3740388e012cf3747bc300d7