Start the `ShareScreenActivity` with the real UserHandle
On large screens, `ShareScreenActivity` is now launched within
the task of the host application that initiated the media
projection. To ensure the `PreShareUI` correctly operates
within this context, the activity must be started with the
host app's real UserHandle.

This change also introduces a NOT_SELECTED screen share mode
to represent the initial state before a user makes a selection
in the large-screen UI.

Flag: com.android.systemui.large_screen_sharing
Bug: 423708479
Test: m on AL
Change-Id: I56ae1b872912f93c5f57db67668bfe736615e8cc