PrintSpooler: Reflect selected color mode in preview
Apps are supposed to check PrintAttributes.getColorMode and render their
printed content in Color or B&W appropriately.  In reality, very few
apps actually do this.  The printer will normally print in monochrome
because the color mode is transmitted with the print job, but the
preview displays in color.  To make the preview more accurately
represent the expected output, re-render the preview pages as grayscale
when B&W is requested.  With this change, neither the app nor the print
service will see any difference in the content, but the user will see a
grayscale preview.  This also requires adding color mode to the
rendering cache key.

Bug: 456450521
Test: atest
Flag: com.android.printspooler.flags.grayscale_preview
Change-Id: I721df1f14d0613d559740d84da0a7ae7e198cbea