[Flexiglass] Remove syntheticScroll updates from NotificationsPlaceholderViewModel
SyntheticScroll updates were originally a way to signal from the nssl to the placeholders, that the bottom-most notification is not visible anymore, and there needs to be some scrolling done to bring it into the viewport again. This feature is broken today, and it is causing more problems by putting the shade into an inconsistent state often.

This CL disables this feature completely to mitigate some more severe issues.

Fixes: 453646977
Bug: 412986215
Test: expand a notification at the bottom of the shade -> no scrolling happens
Flag: com.android.systemui.scene_container
Change-Id: I1443c83d4714c24644d6f29714d70dc799dc72f0