Ensure not to enter desktop in lock task mode
This CL ensures that a task does not enter desktop mode in the lock task
mode.
When the lock task mode is enabled, DesktopTasksController moves a
desktop task out of the desktop mode upon TRANSIT_START_LOCK_TASK_MODE.
But there is a race chance that TO_FRONT can be emitted for the task
before it exits the desktop mode, which ends up re-entering the desktop
mode unexpectedly.
Therefore, this CL prevents such an intermediate transition to move a
task to the desktop while in the transitional state.

Flag: com.android.window.flags.enable_prevent_desktop_in_locktaskmode_bugfix
Bug: 448763429
Test: DesktopTasksControllerTest
Change-Id: I47c77f012be352ba9003ebe59e5cfd4657dd63a4