Remove clearClientSessionsLocked curMethod check
Currently clearClientSessionsLocked checks whether the binding
controller has a bound IME (a non-null IME Invoker), and otherwise skips
most of its work. However, the block that checks it does not have any
dependencies on the IME being bound at all. Moreover, this is always
called before the IME Invoker is cleared in the binding controller.

Previously this method was called clearCurMethodLocked, and it was used
to clear the IME Invoker as well. The name of this method changed in [1]
and the clearing of the IME Invoker was moved to the caller in [2],
making this null check redundant.

With this change we can simply pass in the userId as a parameter,
instead of the whole binding controller.

  [1]: I92c43548a275f358e382a260314bb6fa4546ab42
  [2]: I4b3cbcd1681bae662e644a5c86b5b63df6bf4612

Flag: EXEMPT PURE_REFACTOR
Bug: 281029564
Test: atest CtsInputMethodTestCases
Change-Id: I0db12ddfef84bc77ac7b6eb82dbc1d4b2b8e9b37